+++ scripts/windows/setup.ps1
@@ -1,7 +1,7 @@
 param([string]$GameDir)
 
 $ErrorActionPreference = "Stop"
-$Mgs4Ultra120Version = "v0.3.4-alpha.4"
+$Mgs4Ultra120Version = "v0.3.4-alpha.5"
 Add-Type -AssemblyName System.Windows.Forms
 Add-Type -AssemblyName System.Drawing
 [Windows.Forms.Application]::EnableVisualStyles()